requirements.txt
torch
torchvision
faiss-cpu
pandas
numpy
matplotlib
seaborn
scikit-learn
tqdm
opencv-python
Pillow
streamlit
pillow
